+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -7,8 +7,8 @@ import java.util.*;
 
 @Component
 public class InMemoryUserStorage implements UserStorage {
-    private final Map<Long, User> users = new HashMap<>();
-    private long currentId = 1;
+    private final Map<Integer, User> users = new HashMap<>();
+    private Integer currentId = 1;
 
     @Override
     public User addUser(User user) {
@@ -27,7 +27,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public User getUser(long id) {
+    public User getUser(Integer id) {
         return users.get(id);
     }
